% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_reviews.R
\name{filter_reviews}
\alias{filter_reviews}
\title{Filter reviews}
\usage{
filter_reviews(data)
}
\arguments{
\item{data}{Reactive function brings back all the Comments with selected rating}
}
\value{
Brings back all the comments with selected rating
}
\description{
Filter reviews
}
